refacor: move tests next together
