+++ src/builder/JsonSchemaBuilder.ts
@@ -58,56 +58,6 @@ export class JsonSchemaBuilder {
     }
   }
 
-  //   private deleteNestedPropertyByPath = (obj: JsonSchema, path: string): JsonSchema => {
-  //     const keys = path.split('.');
-  //     const newObject = { ...obj };
-  //
-  //     if (keys.length === 0) {
-  //       return newObject;
-  //     }
-  //
-  //     let current = newObject;
-  //     const stack = [];
-  //
-  //     for (let i = 0; i < keys.length - 1; i++) {
-  //       stack.push(current); //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //       current[keys[i]] = { ...current[keys[i]] }; //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //       current = current[keys[i]];
-  //     }
-  // //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //     delete current[keys[keys.length - 1]];
-  //
-  //     for (let i = keys.length - 2; i >= 0; i--) {
-  //       const key = keys[i]; //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //
-  //       current = stack.pop(); //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //       if (Object.keys(current[key]).length === 0) {//@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //         delete current[key];
-  //       }
-  //     }
-  //
-  //     return newObject;
-  //   };
-  //
-  //   private updateNestedObjectByPath = (obj: JsonSchema, path: string, value: unknown): JsonSchema => {
-  //     console.log('🐕 sag path', path); // TODO: REMOVE ME ⚠️
-  //
-  //     const keys = path.split('.');
-  //     const newObject = { ...obj };
-  //
-  //     let current = newObject;
-  //     keys.forEach((key, index) => {
-  //       if (index === keys.length - 1) { //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //         current[key] = value;
-  //       } else { //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //         current[key] = current[key] ? { ...current[key] } : {}; //@ts-expect-error // eslint-disable-line @typescript-eslint/ban-ts-comment
-  //         current = current[key];
-  //       }
-  //     });
-  //
-  //     return newObject;
-  //   };
-
   private deleteNestedPropertyByPath = (obj: JsonSchema, path: string): JsonSchema => {
     const keys = path.split('.');
     const newObject = { ...obj };
